+++ Music/Content.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 
+#if DEBUG
 public var musicLibrary = [
     URL.init(fileURLWithPath: Bundle.main.path(forResource: "Billie Eilish - Bad Guy", ofType: "mp3")!),
     URL.init(fileURLWithPath: Bundle.main.path(forResource: "Ed Sheeran - Perfect", ofType: "mp3")!),
@@ -14,3 +15,11 @@ public var musicLibrary = [
     URL.init(fileURLWithPath: Bundle.main.path(forResource: "Katy Perry - I Kissed A Girl", ofType: "mp3")!),
     URL.init(fileURLWithPath: Bundle.main.path(forResource: "OneRepublic & SeeB - Rich Love", ofType: "mp3")!)
 ]
+#else
+public var musicLibrary = [
+    URL.init(fileURLWithPath: Bundle.main.path(forResource: "Oddvision-Media-Infraction-Music-Royal-Stage-water", ofType: "mp3")!),
+    URL.init(fileURLWithPath: Bundle.main.path(forResource: "Infraction-Let-Me-Know-Main-Version-pr", ofType: "mp3")!),
+    URL.init(fileURLWithPath: Bundle.main.path(forResource: "Infraction-Music-Take-A-Break-pr", ofType: "mp3")!),
+    URL.init(fileURLWithPath: Bundle.main.path(forResource: "Infraction-Almost-Evil-pr", ofType: "mp3")!)
+]
+#endif
