+++ Music/Extensions.swift
@@ -12,14 +12,20 @@ import UIKit
 struct Song {
     var title: String
     var artist: String
-    var artwork: UIImage?
+    var artwork: UIImage
 }
 
+let defaultArtwork = UIImage(named: "DefaultThumbnail")!
+
+let playImage = UIImage(systemName: "play.fill")?.withConfiguration(UIImage.SymbolConfiguration(pointSize: 50, weight: .regular))
+let pauseImage = UIImage(systemName: "pause.fill")?.withConfiguration(UIImage.SymbolConfiguration(pointSize: 50, weight: .regular))
+
+// Parsing mp3 file to get song data
 func getSong(url: URL) async throws -> Song {
     let asset = AVAsset(url: url)
     var songName: String = "Unknown"
     var artistName: String = "Unknown"
-    var artworkImage: UIImage?
+    var artworkImage: UIImage = defaultArtwork
     
     for format in try await asset.load(.availableMetadataFormats) {
         let metadata = try await asset.loadMetadata(for: format)
@@ -37,7 +43,7 @@ func getSong(url: URL) async throws -> Song {
                     }
                 case "artwork":
                     if let artworkData = try await item.load(.dataValue) {
-                        artworkImage = UIImage(data: artworkData)
+                        artworkImage = UIImage(data: artworkData) ?? defaultArtwork
                     }
                 default:
                     break
@@ -46,38 +52,5 @@ func getSong(url: URL) async throws -> Song {
         }
     }
 
-    print("Song name: \(songName), Artist: \(artistName), Artwork: \(artworkImage)")
     return Song(title: songName, artist: artistName, artwork: artworkImage)
 }
-/*
-func getSong(url: URL) async throws -> Song {
-    let asset = AVAsset(url: url)
-    var songTitle: String
-    var artistName: String
-    
-    for format in try await asset.load(.availableMetadataFormats) {
-        let metadata = try await asset.loadMetadata(for: format)
-        // Process the format-specific metadata collection.
-        for item in metadata {
-            if let commonKey = item.commonKey?.rawValue {
-                switch commonKey {
-                    case "title":
-                        if let title = item.stringValue {
-                            songTitle = title
-                        }
-                    case "artist":
-                        if let artist = item.stringValue {
-                            artistName = artist
-                        }
-                    default:
-                        break
-            }
-                print("Metadata key: \(commonKey)")
-            }
-        }
-    }
-    // Example song, should extract actual song name and artist from metadata
-    return Song(title: "Perfect", artist: "Ed Sheeran")
-}
-
-*/
